fix: sqlite write contention
in the sqlite backend, the current strategy for writes is to
write-with-retry. however, this can fail because our connection could be
starved for a write if we are racing with other processes and sqlite
keeps giving the write to other processes.

this commit addresses the issue by acquiring an exclusive lock from
sqlite before any write to the database.

references:
- https://stackoverflow.com/a/21888761
- https://www.sqlite.org/lockingv3.html

update tests