feat: make table creation optional (#1097)
* feat: make table creation optional

This is a backwards compatible change that makes dynamodb table creation optional. Will high levels of load this was leading to API throttling errors for the dynamodb create table api. Solves https://github.com/requests-cache/requests-cache/issues/1027

* docs: update the docs, docstrings, and changelog

* fix: trailing whitespace

* test: add tests

Add a test case which ensures that the case of _not_ creating a table is
covered.

* fix: formatting

* fix: use plain assertion